+++ components/OpenCageSearch/index.js
@@ -3,7 +3,7 @@ import style from './styles.sass'
 
 import { h, Component } from 'preact'
 import { PropTypes } from 'preact-compat'
-import { autocomplete, fixBerlinSearchResult } from '../../lib/opencage.js'
+import { autocomplete, fixBerlinSearchResult } from '../../lib/openrouteservice.js'
 import searchIcon from './img/searchIcon.svg'
 import closeIcon from './img/closeIcon.svg'
